change SimpleDateFormat to DateTimeFormatter and move formatter to companion object

+++ app/src/main/java/com/daggery/nots/data/Note.kt
@@ -11,5 +11,5 @@ data class Note(
     @ColumnInfo(name = "note_order") val noteOrder: Int,
     @ColumnInfo(name = "note_title") val noteTitle: String,
     @ColumnInfo(name = "note_body") val noteBody: String,
-    @ColumnInfo(name = "note_date") val noteDate: Int,
+    @ColumnInfo(name = "note_date") val noteDate: Long,
 )
